DEV: properly fix 13161 (#950)

+++ application/helpers/expressions/em_manager_helper.php
@@ -5543,9 +5543,9 @@
                         $criteria->addCondition('srid=:srid');
                         $criteria->addCondition('sid=:sid');
                         $criteria->params = [':srid'=>$_SESSION[$this->sessid]['srid'],':sid'=>$this->sid];
-                        $savedControls = SavedControl::model()->findAll($criteria);
+                        $savedControl = SavedControl::model()->find($criteria);
 
-                        foreach ($savedControls as $savedControl) {
+                        if($savedControl){
                             $savedControl->delete();
                         }
